# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
  isAdmin?: boolean;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface SignupRequest {
  email: string;
  /** @minLength 6 */
  password: string;
  firstName?: string;
  lastName?: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

export interface Product {
  id: number;
  name: string;
  description: string;
  price: number;
  imageUrl: string;
  category: string;
  inStock: string;
  createdAt?: string;
}

export interface ProductInput {
  name: string;
  description?: string;
  price: number;
  imageUrl: string;
  category: string;
  inStock?: string;
}

export interface DeletedResponse {
  success: boolean;
}

export type OrderInputItemsItem = { [key: string]: unknown };

export interface OrderInput {
  razorpayPaymentId: string;
  razorpayOrderId: string;
  totalAmount: number;
  email?: string;
  contact?: string;
  items: OrderInputItemsItem[];
}

export type OrderSavedItemsItem = { [key: string]: unknown };

export interface OrderSaved {
  id: number;
  razorpayPaymentId?: string;
  email?: string;
  contact?: string;
  totalAmount?: number;
  status?: string;
  createdAt?: string;
  items?: OrderSavedItemsItem[];
}

/**
 * Opaque session token — Bearer <sid>.
 */
export type AuthorizationSessionHeaderParameter = string;

export type ListProductsParams = {
category?: string;
};