# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: Auth
    description: Browser and mobile authentication endpoints.
  - name: products
    description: Product catalog
  - name: orders
    description: Order management
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /auth/user:
    get:
      tags: [Auth]
      operationId: getCurrentAuthUser
      summary: Get the currently authenticated user
      parameters:
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: Auth status resolved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserEnvelope'
  /auth/signup:
    post:
      tags: [Auth]
      operationId: signupWithEmail
      summary: Create account with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupRequest'
      responses:
        '201':
          description: Account created and session started.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserEnvelope'
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Email already registered.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /auth/login:
    post:
      tags: [Auth]
      operationId: loginWithEmail
      summary: Sign in with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authenticated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserEnvelope'
        '401':
          description: Invalid credentials.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /auth/logout:
    post:
      tags: [Auth]
      operationId: logoutSession
      summary: Clear the session cookie
      responses:
        '200':
          description: Logged out.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutSuccess'
  /products:
    get:
      tags: [products]
      operationId: listProducts
      summary: List all products
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Product list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
    post:
      tags: [products]
      operationId: createProduct
      summary: Create a product (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
  /products/{id}:
    put:
      tags: [products]
      operationId: updateProduct
      summary: Update a product (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductInput'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
    delete:
      tags: [products]
      operationId: deleteProduct
      summary: Delete a product (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeletedResponse'
  /orders:
    post:
      tags: [orders]
      operationId: createOrder
      summary: Save a completed order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderInput'
      responses:
        '201':
          description: Order saved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderSaved'
    get:
      tags: [orders]
      operationId: listOrders
      summary: List all orders (admin only)
      responses:
        '200':
          description: Order list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderSaved'
  /orders/{id}/ship:
    patch:
      tags: [orders]
      operationId: shipOrder
      summary: Mark an order as shipped (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Order updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderSaved'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      description: Opaque session token — Bearer <sid>.
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie.
      schema:
        type: string
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    AuthUser:
      type: object
      required: [id, email, firstName, lastName, profileImageUrl]
      properties:
        id:
          type: string
        email:
          type: [string, 'null']
          format: email
        firstName:
          type: [string, 'null']
        lastName:
          type: [string, 'null']
        profileImageUrl:
          type: [string, 'null']
        isAdmin:
          type: boolean
    AuthUserEnvelope:
      type: object
      required: [user]
      properties:
        user:
          oneOf:
            - $ref: '#/components/schemas/AuthUser'
            - type: 'null'
    SignupRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        firstName:
          type: string
        lastName:
          type: string
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    LogoutSuccess:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          const: true
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
    Product:
      type: object
      required: [id, name, description, price, imageUrl, category, inStock]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: integer
        imageUrl:
          type: string
        category:
          type: string
        inStock:
          type: string
        createdAt:
          type: string
    ProductInput:
      type: object
      required: [name, price, imageUrl, category]
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: integer
        imageUrl:
          type: string
        category:
          type: string
        inStock:
          type: string
    DeletedResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
    OrderInput:
      type: object
      required: [razorpayPaymentId, razorpayOrderId, totalAmount, items]
      properties:
        razorpayPaymentId:
          type: string
        razorpayOrderId:
          type: string
        totalAmount:
          type: integer
        email:
          type: string
        contact:
          type: string
        items:
          type: array
          items:
            type: object
    OrderSaved:
      type: object
      required: [id]
      properties:
        id:
          type: integer
        razorpayPaymentId:
          type: string
        email:
          type: string
        contact:
          type: string
        totalAmount:
          type: integer
        status:
          type: string
        createdAt:
          type: string
        items:
          type: array
          items:
            type: object