# lib/api-zod/src/generated/types/orderSaved.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderSavedItemsItem } from './orderSavedItemsItem';

export interface OrderSaved {
  id: number;
  razorpayPaymentId?: string;
  email?: string;
  contact?: string;
  totalAmount?: number;
  status?: string;
  createdAt?: string;
  items?: OrderSavedItemsItem[];
}